+++ source/renderer/app/components/staking/delegation-center/UndelegateConfirmationResultDialog.js
@@ -3,6 +3,7 @@ import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import { defineMessages, intlShape, FormattedHTMLMessage } from 'react-intl';
 import SVGInline from 'react-svg-inline';
+import humanizeDuration from 'humanize-duration';
 import DialogCloseButton from '../../widgets/DialogCloseButton';
 import Dialog from '../../widgets/Dialog';
 import styles from './UndelegateConfirmationResultDialog.scss';
@@ -18,12 +19,17 @@ const messages = defineMessages({
   description: {
     id: 'staking.delegationCenter.undelegate.result.dialog.description',
     defaultMessage:
-      '!!!<p>The stake from your wallet <strong>{walletName}</strong> is no longer delegated and you will soon stop earning rewards for this wallet.</p><p>Your new delegation preferences are now posted on the blockchain <strong>and will take effect at the start of the next Cardano epoch in 3 hours and 14 minutes</strong>. For the rest of the current epoch, your previous delegation preferences are still active.</p>',
+      '!!!<p>The stake from your wallet <strong>{walletName}</strong> is no longer delegated and you will soon stop earning rewards for this wallet.</p><p>Your new delegation preferences are now posted on the blockchain <strong>and will take effect at the start of the next Cardano epoch in {timeUntilNextEpochStart}</strong>. For the rest of the current epoch, your previous delegation preferences are still active.</p>',
     description: 'Description for the "Undelegate Result" dialog.',
   },
 });
 
-type Props = { walletName: string, onClose: Function };
+type Props = {
+  walletName: string,
+  nextEpochStartTime: string,
+  currentLocale: string,
+  onClose: Function,
+};
 
 @observer
 export default class UndelegateConfirmationResultDialog extends Component<Props> {
@@ -33,7 +39,7 @@ export default class UndelegateConfirmationResultDialog extends Component<Props>
 
   render() {
     const { intl } = this.context;
-    const { walletName, onClose } = this.props;
+    const { walletName, onClose, nextEpochStartTime, currentLocale } = this.props;
     const actions = [
       {
         label: intl.formatMessage(globalMessages.close),
@@ -41,6 +47,39 @@ export default class UndelegateConfirmationResultDialog extends Component<Props>
         primary: true,
       },
     ];
+
+    const timeLeft = Math.max(
+      0,
+      new Date(nextEpochStartTime).getTime() - new Date().getTime()
+    );
+
+    let humanizedDurationLanguage;
+    switch (currentLocale) {
+      case 'ja-JP':
+        humanizedDurationLanguage = 'ja';
+        break;
+      case 'zh-CN':
+        humanizedDurationLanguage = 'zh_CN';
+        break;
+      case 'ko-KR':
+        humanizedDurationLanguage = 'ko';
+        break;
+      case 'de-DE':
+        humanizedDurationLanguage = 'de';
+        break;
+      default:
+        humanizedDurationLanguage = 'en';
+    }
+
+    const timeUntilNextEpochStart = humanizeDuration(timeLeft || 0, {
+      round: true, // round seconds to prevent e.g. 1 day 3 hours *11,56 seconds*
+      language: humanizedDurationLanguage,
+      conjunction: ' and ',
+      units: ['d', 'h', 'm'],
+      serialComma: false,
+    });
+
+    console.debug('timeUntilNextEpochStart: ', timeUntilNextEpochStart);
 
     return (
       <Dialog
@@ -57,7 +96,10 @@ export default class UndelegateConfirmationResultDialog extends Component<Props>
         <div className={styles.description}>
           <FormattedHTMLMessage
             {...messages.description}
-            values={{ walletName }}
+            values={{
+              walletName,
+              timeUntilNextEpochStart,
+            }}
           />
         </div>
       </Dialog>