+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -127,11 +127,9 @@ export const DARK_BLUE_THEME_OUTPUT = {
   },
   delegationSetupWizard: {
     '--theme-delegation-steps-activation-steps-indicator-color': '#e9f4fe',
-    '--theme-delegation-steps-activation-description-color':
-      'rgba(233, 244, 254, 0.8)',
-    '--theme-delegation-steps-activation-fees-label-color': '#e9f4fe',
-    '--theme-delegation-steps-activation-fees-amount-color': '#ea4c5b',
-    '--theme-delegation-steps-activation-address-value-color': '#e9f4fe',
+    '--theme-delegation-steps-success-description-color':
+      'rgba(233, 244, 254, 1)',
+    '--theme-delegation-steps-success-tada-color': 'rgba(233, 244, 254, 1)',
     '--theme-delegation-steps-choose-stake-pool-checkmark-icon-color':
       '#536370',
     '--theme-delegation-steps-choose-stake-pool-delegated-pools-label-color':
@@ -148,7 +146,7 @@ export const DARK_BLUE_THEME_OUTPUT = {
     '--theme-delegation-steps-choose-stake-pool-thumb-border-color':
       'rgba(102, 122, 138, 0.2)',
     '--theme-delegation-steps-choose-stake-pool-title-color':
-      'rgba(233, 244, 254, 0.8)',
+      'rgba(233, 244, 254, 1)',
     '--theme-delegation-steps-choose-stake-pool-tooltip-arrow-color':
       'rgba(94, 96, 102, 0.9)',
     '--theme-delegation-steps-choose-stake-pool-tooltip-background-color':