+++ storybook/stories/staking/Staking.stories.js
@@ -138,6 +138,13 @@ storiesOf('Decentralization | Staking', module)
   .add('Delegation Wizard', props => <StakingDelegationSteps {...props} />, {
     id: 'wizard',
   })
+  .add(
+    'Delegation Wizard - Delegation Not Available',
+    props => <StakingDelegationSteps {...props} isDisabled />,
+    {
+      id: 'wizard',
+    }
+  )
   .add('Undelegate Confirmation', StakingUndelegateConfirmationStory, {
     id: 'undelegate-confirmation',
   })